git: Add new aliases for status and branch, remove from bash
Add `git s` and `git b` aliases, but remove the bash aliases `gits` and
`gitb`. This is because the git autocomplete obviously doesn't expand
bash aliases, and adding an extra space to my muscle memory is easier
than making me type `git st` and `git br` or whatever the standard it.