Doom: Remove 'vaa' binding, Vim: 'vaa'->'vag'
My 'vaa' binding was breaking visual mode, so instead of making it an
explicit binding, I made a 'whole-buffer' text object and bound it to
"a", this way I could use the same binding. Turns out, however, Doom
already binds this text object to 'g'.

Let's use the same binding in Vim.