Neovim: Add helper to allow setup before install
With the old way we try to require a module and then call setup on it.
The problem is that if we add a new module and add the setup invocation,
the next time we open Vim there's be errors since we are trying to
require the module before we've installed it.

This helper makes it so that we just fail silently if we try to require
a module that hasn't been installed yet.